pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "neoswga"
version = "3.0.0"
description = "Enhanced Selective Whole Genome Amplification (SWGA) primer design with advanced thermodynamics and optimization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Andreas Sjodin", email = "andreas.sjodin@example.com"},
]
maintainers = [
    {name = "Andreas Sjodin", email = "andreas.sjodin@example.com"},
]
keywords = [
    "bioinformatics",
    "genomics",
    "primer-design",
    "SWGA",
    "whole-genome-amplification",
    "thermodynamics",
    "genetic-algorithm",
    "machine-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.18.0",
    "pandas>=1.0.0",
    "scipy>=1.4.1",
    "scikit-learn>=0.21.3",
    "h5py>=2.10.0",
    "networkx>=2.5",
    "biopython>=1.78",
    "joblib>=0.14.1",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "melt>=1.0.3",
    "openpyxl>=3.0.0",
]

[project.optional-dependencies]
improved = [
    "pybloom-live>=3.1.0",
    "mip>=1.13.0",
]
gpu = [
    "cupy-cuda11x>=11.0.0",
]
gpu-cuda12 = [
    "cupy-cuda12x>=12.0.0",
]
deep-learning-torch = [
    "torch>=1.10.0",
]
deep-learning-tf = [
    "tensorflow>=2.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
]
all = [
    "pybloom-live>=3.1.0",
    "mip>=1.13.0",
    "cupy-cuda11x>=11.0.0",
    "torch>=1.10.0",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/neoswga"
Documentation = "https://github.com/yourusername/neoswga/blob/master/README.md"
Repository = "https://github.com/yourusername/neoswga"
"Bug Tracker" = "https://github.com/yourusername/neoswga/issues"

[project.scripts]
neoswga = "neoswga.cli_unified:main"

[tool.setuptools]
packages = ["neoswga", "neoswga.core"]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.package-data]
neoswga = ["*.md"]
"neoswga.core" = ["*.md", "models/*.p"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=neoswga",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
